Har skapat två packet controller och model. Samt två klasser Product med constuctor, get & set, och ProductsController.
Har ändrat BigDecimal till flot.
Fyllt i ProductsController med produkter för att testa att det syns.
Jag hade problem med denna kod tidigare så det är fel någonstans.

+++ src/main/java/se/iths/cecilia/usersandproducts/model/Product.java
@@ -1,18 +1,17 @@
 package se.iths.cecilia.usersandproducts.model;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class Product {
     private BigInteger id;
     private String title;
-    private BigDecimal price;
+    private float price;
     private String name;
 
     public Product() {
     }
 
-    public Product(String title, BigDecimal price, String name) {
+    public Product(String title, float price, String name) {
         this.title = title;
         this.price = price;
         this.name = name;
@@ -34,11 +33,11 @@ public class Product {
         this.title = title;
     }
 
-    public BigDecimal getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(float price) {
         this.price = price;
     }
